Refactor database logic for possible diesel upgrade
