+++ Makefile.toml
@@ -184,6 +184,12 @@ dependencies = [
 command = "cargo"
 args = ["test", "--workspace", "--no-default-features", "--features=rust-fuzz", "rust_fuzz", "--", "--skip", "generated"]
 
+[tasks.test-db-diesel-all]
+dependencies = [
+    "test-db-diesel-mysql",
+    "test-db-diesel-postgres",
+]
+
 [tasks.test-db-mysql-all]
 dependencies = [
     "test-db-diesel-mysql"
@@ -191,7 +197,7 @@ dependencies = [
 
 [tasks.test-db-diesel-mysql]
 command = "cargo"
-args = ["test", "--workspace", "--tests", "--features=db-diesel-mysql", "db", "--", "--skip", "generated"]
+args = ["test", "--workspace", "--tests", "--features=db-diesel-mysql", "mysql", "--", "--skip", "generated"]
 
 [tasks.test-db-postgres-all]
 dependencies = [
@@ -202,15 +208,15 @@ dependencies = [
 
 [tasks.test-db-postgres]
 command = "cargo"
-args = ["test", "--workspace", "--tests", "--features=db-postgres", "db", "--", "--skip", "generated"]
+args = ["test", "--workspace", "--tests", "--features=db-postgres", "postgres", "--", "--skip", "generated"]
 
 [tasks.test-db-tokio-postgres]
 command = "cargo"
-args = ["test", "--workspace", "--tests", "--features=db-tokio-postgres", "db", "--", "--skip", "generated"]
+args = ["test", "--workspace", "--tests", "--features=db-tokio-postgres", "postgres", "--", "--skip", "generated"]
 
 [tasks.test-db-diesel-postgres]
 command = "cargo"
-args = ["test", "--workspace", "--tests", "--features=db-diesel-postgres", "db", "--", "--skip", "generated"]
+args = ["test", "--workspace", "--tests", "--features=db-diesel-postgres", "postgres", "--", "--skip", "generated"]
 
 [tasks.test-rocket-traits]
 command = "cargo"
